+++ bbootimg/src/main/kotlin/avb/desc/ChainPartitionDescriptor.kt
@@ -8,22 +8,22 @@ import java.util.*
 
 @OptIn(ExperimentalUnsignedTypes::class)
 class ChainPartitionDescriptor(
-        var rollback_index_location: UInt = 0U,
-        var partition_name_len: UInt = 0U,
-        var public_key_len: UInt = 0U,
+        var rollback_index_location: Int = 0,
+        var partition_name_len: Int = 0,
+        var public_key_len: Int = 0,
         var partition_name: String = "",
         var pubkey: ByteArray = byteArrayOf(),
         var pubkey_sha1: String = ""
-) : Descriptor(TAG, 0U, 0) {
+) : Descriptor(TAG, 0, 0) {
     override fun encode(): ByteArray {
-        this.partition_name_len = this.partition_name.length.toUInt()
-        this.public_key_len = this.pubkey.size.toUInt()
-        this.num_bytes_following = (SIZE.toUInt() + this.partition_name_len + this.public_key_len - 16U).toULong()
-        val nbf_with_padding = Helper.round_to_multiple(this.num_bytes_following.toLong(), 8).toULong()
-        val padding_size = nbf_with_padding - this.num_bytes_following
+        this.partition_name_len = this.partition_name.length
+        this.public_key_len = this.pubkey.size
+        this.num_bytes_following = SIZE + this.partition_name_len + this.public_key_len - 16
+        val nbf_with_padding = Helper.round_to_multiple(this.num_bytes_following, 8).toULong()
+        val padding_size = nbf_with_padding - this.num_bytes_following.toUInt()
         val desc = Struct3(FORMAT_STRING + "${RESERVED}x").pack(
                 TAG,
-                nbf_with_padding.toULong(),
+                nbf_with_padding,
                 this.rollback_index_location,
                 this.partition_name.length.toUInt(),
                 this.public_key_len,
@@ -33,7 +33,7 @@ class ChainPartitionDescriptor(
     }
 
     companion object {
-        const val TAG: ULong = 4U
+        const val TAG: Long = 4L
         const val RESERVED = 64
         const val SIZE = 28L + RESERVED
         const val FORMAT_STRING = "!2Q3L"
@@ -45,14 +45,14 @@ class ChainPartitionDescriptor(
         }
         this.sequence = seq
         val info = Struct3(FORMAT_STRING + "${RESERVED}s").unpack(data)
-        this.tag = info[0] as ULong
-        this.num_bytes_following = info[1] as ULong
-        this.rollback_index_location = info[2] as UInt
-        this.partition_name_len = info[3] as UInt
-        this.public_key_len = info[4] as UInt
+        this.tag = (info[0] as ULong).toLong()
+        this.num_bytes_following = (info[1] as ULong).toLong()
+        this.rollback_index_location = (info[2] as UInt).toInt()
+        this.partition_name_len = (info[3] as UInt).toInt()
+        this.public_key_len = (info[4] as UInt).toInt()
         val expectedSize = Helper.round_to_multiple(
-                SIZE.toUInt() - 16U + this.partition_name_len + this.public_key_len, 8U)
-        if (this.tag != TAG || this.num_bytes_following != expectedSize.toULong()) {
+                SIZE - 16 + this.partition_name_len + this.public_key_len, 8)
+        if (this.tag != TAG || this.num_bytes_following != expectedSize.toLong()) {
             throw IllegalArgumentException("Given data does not look like a chain/delegation descriptor")
         }
         val info2 = Struct3("${this.partition_name_len}s${this.public_key_len}b").unpack(data)
