+++ bbootimg/src/main/kotlin/avb/desc/KernelCmdlineDescriptor.kt
@@ -9,7 +9,7 @@ class KernelCmdlineDescriptor(
         var flags: UInt = 0U,
         var cmdlineLength: UInt = 0U,
         var cmdline: String = "")
-    : Descriptor(TAG, 0U, 0) {
+    : Descriptor(TAG, 0, 0) {
     var flagsInterpretation: String = ""
         get() {
             var ret = ""
@@ -24,13 +24,13 @@ class KernelCmdlineDescriptor(
     @Throws(IllegalArgumentException::class)
     constructor(data: InputStream, seq: Int = 0) : this() {
         val info = Struct3(FORMAT_STRING).unpack(data)
-        this.tag = info[0] as ULong
-        this.num_bytes_following = info[1] as ULong
+        this.tag = (info[0] as ULong).toLong()
+        this.num_bytes_following = (info[1] as ULong).toLong()
         this.flags = info[2] as UInt
         this.cmdlineLength = info[3] as UInt
         this.sequence = seq
         val expectedSize = Helper.round_to_multiple(SIZE.toUInt() - 16U + this.cmdlineLength, 8U)
-        if ((this.tag != TAG) || (this.num_bytes_following != expectedSize.toULong())) {
+        if ((this.tag != TAG) || (this.num_bytes_following != expectedSize.toLong())) {
             throw IllegalArgumentException("Given data does not look like a kernel cmdline descriptor")
         }
         this.cmdline = Struct3("${this.cmdlineLength}s").unpack(data)[0] as String
@@ -50,7 +50,7 @@ class KernelCmdlineDescriptor(
     }
 
     companion object {
-        const val TAG: ULong = 3U
+        const val TAG: Long = 3L
         const val SIZE = 24
         const val FORMAT_STRING = "!2Q2L" //# tag, num_bytes_following (descriptor header), flags, cmdline length (bytes)
         //AVB_KERNEL_CMDLINE_FLAGS_USE_ONLY_IF_HASHTREE_NOT_DISABLED
